MLK-11316-1 mxc IPUv3: PRE: Export a function to set PRE_CTRL register
In order to workaround the PRE SoC bug recorded by errata ERR009624, the
software cannot write the PRE_CTRL register when the PRE writes the PRE_CTRL
register automatically to set the ENABLE bit(bit0) to 1 in the PRE repeat mode.

This patch exports a function to set the PRE_CTRL register so that it could be
used by the software when the PRE automatic writing doesn't happen for sure.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>