ARM: dts: ccimx6qsbc: zero gpio-reset flag for Ethernet PHY
The Ethernet PHY used gpio flag 2 for the reset line.
This is wrong because the i.MX6 gpio driver only admits
flags 0|1 for polarity.
This change however shouldn't have any functional impact
because the i.MX6 Ethernet driver doesn't really care
about the flag and hardcodes the toggling of this GPIO
assuming it is active low.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>