ARM: dts: ccimx6: Do not specify suspend voltages when not strictly needed.
The policy for the behaviour of regulators on suspend is as follows:

1) When the regulator is used in a driver, the suspend function in the
   driver should configure it for suspend, even disabling it if not needed.
2) If the regulator is shared by multiple drivers but we still want to
   configure it (or disable it) on suspend, we can do this using the
   "regulator-suspend*" device tree properties.

Regulators that define the suspend and standby voltage settings to be the
same as the normal "A" settings do not require to change voltages on
suspend.

On the v2015.04 u-boot, i.mx6q 1.2GHz variants now run at a higher
voltage than 1GHz or 800MHz variants, so we would need to have
different device trees if we wanted to specify the suspend/standby
voltages.

So, this commit removes the unneeded suspend configuration for those
regulators that were not changing the voltage or disabling the regulator
on suspend, to avoid having more device tree files for the 1.2GHz variants.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>