ARM: defconfig: ccimx6sbc: Disable built-in GPU driver.
DEY-2.0 compiles its own GPU driver as an external module. This means that
the built-in module may be out of sync with user space and will not work.

This change has already been done in the DEY defconfig:
8016bd84e9e9db4a521b62247096c63b4b0cda6a

At this moment the built-in driver does not work with the user space.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>