Revert "ASoC: imx-sgtl5000: enable DAC only when headphone is connected"
This reverts commit 8ad126409b79a445f6afc7e92dc68652592d7680.
Although this patch seemed to minimize the impact of the issue, it
didn't completely cancel it, and the work around had other issues,
like the fact that the audio is cancelled on the LINE-OUT connector
too, if the headphone jack is disconnected.

The original issue has been addressed by hardware redesign of the SBC
and is fixed on SBCv2 and further versions.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2128
https://jira.digi.com/browse/DEL-1572