arm: dts: ccimx6sbc: disable additional interfaces on DTSI include file
In previous commits:
	19426ac950f8
	e8ffa6e25f18
	7fa5b4cda4fa
the property 'status="okay"' was moved from DTSI to the platforms' DTS
files but for SBC only devices these nodes are not specifically disabled
in a previous included DTSI so the omission of the 'status' property
defaults to "okay" (enabled).
Disable these devices specifically in the DTSI include file since they
are enabled by the corresponding platform DTS files.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2162https://jira.digi.com/browse/DEL-2162