drivers: da9063-onkey: report SLEEP only if the system is not suspended
On Digi Embedded Yocto, when the system is suspended, resuming via the
power button reports SLEEP to the ACPI process which triggers the suspend
procedure racing with the resuming operation. This may result in the system
going back to suspend when resuming from the power button.

This patch avoids that problem reporting SLEEP only when the driver has been
resumed

https://jira.digi.com/browse/DEL-2117

Signed-off-by: Jose Diaz de Grenu de Pedro <Jose.DiazdeGrenudePedro@digi.com>