arm: dts: ccimx6sbc: enable card detect polling on uSD
The uSD card controller on the SBC had set the property
'non-removable' to allow the interface to suspend/resume reliably
when the system was booted from the uSD. This however prevented
the system from doing a polling for card removal/insertion.

The uSD is unlikely to be used for booting the system, but
rather be used for hotplug of cards. Remove the non-removable
property to allow the card detection polling to work.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2213