mxc: ldb: move notifier block to each LDB channel
There is one notifier block for the LDB which was used to show the logo
on the displays. Since there may be up to two LVDS channels working at
the same time, we need to have one notifier block per channel. Otherwise,
when enabling the two LVDS channels the kernel hangs.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-1901