ccimx6ul_defconfig: Do not enable bluetooth sdio for qca6564
The qca6564 chip has some SDIO conflict with the BT SDIO driver so when
both are present the kernel uses the BT SDIO instead of the MMC SDIO.
If BT_HCIUART_IBS is present it means the kernel has been compiled for the
qca6564 so add a dependency to do not enable BT_HCIBTSDIO if BT_HCIUART_IBS
is enabled.
Additionally also enable the specific qca6564 driver.

https://jira.digi.com/browse/DEL-2581

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>