ARM: dts: imx6ul-ccimx6ul: modify ldo3 pmic entry
The SOM revision 3 does not use the ldo3 for powering the NAND, so rename
that entry. Additionally, that regulator is enabled in the PMIC OTP
configuration but is not currently connected, so we can disable it on suspend.

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>
(cherry picked from commit 5cc159de7c0b63b33550df6ccd0a659d2c53f0e6)