net: fec: free PHY reset GPIO to be able to do a second hardware reset
The fec_reset_phy() function requests the PHY reset GPIO and never frees
it. This prevents the driver from calling the reset function a second time
since it won't be able to request the GPIO again.

This commit adds a flag to free the GPIO if the reset function needs to be
called another time.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2701