mfd: mca: gpio: save gpio_base in mca structure
This commit implements the code to save in the main mca
structure the gpio_base of the mca-gpios. This means, the
base number of the GPIOs handled by the mca-gpio driver.

This allows other drivers to use that GPIOs, without the
need of passing that parameter through the device-tree.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit 0117f1bb4c6f9b8d6bdcdad223813d3912c48d6d)