ARM: imx6ql: ov5640_mipi: Rename compatible property to match driver.
There are now two versions of the ov5640_mipi driver, one for the v4l2
int-device architecture and one for the v4l2 subdev architecture.

This commits renames the device tree node to match the one for the int-device
architecture that changed name in the driver when the subdev driver was added.

https://jira.digi.com/browse/DEL-2867

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>