mxc: capture: ov564x: Only exit ov564x_reset() if neither pwd or rst GPIOs are available.
the reset() function is keeping one of the ov564x cameras on reset while
resetting the I2C client address of the other.

This is done to be able to use a dual camera setup when both sensors
have the same default client I2C address initially.

If neither a power down or a reset lines are provided, the cameras cannot
be kept on reset so there is no point in trying to update the address.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
Signed-off-by: Hector Palacios <hector.palacios@digi.com>

(cherry picked from commit d896181bc18cb9cb5f5ccfde6b1da6ac007ed78d)
Signed-off-by: Javier Viguera <javier.viguera@digi.com>