mxc: capture: ov564x: Only exit ov564x_reset() if neither pwd or rst GPIOs are available.
the reset() function is keeping one of the ov564x cameras on reset while
resetting the I2C client address of the other.

This is done to be able to use a dual camera setup when both sensors
have the same default client I2C address initially.

If neither a power down or a reset lines are provided, the cameras cannot
be kept on reset so there is no point in trying to update the address.

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit f3bf547e9f7445481625ce26c3ea374d4302d255)
Signed-off-by: Hector Palacios <hector.palacios@digi.com>