net: fec: fix free of FEC's PHY reset GPIO line
Commit 546935983421 introduced a static variable to flag whether the PHY
reset line had been requested and could be freed in order to whack another
reset. This mechanism does not work well for supporting two FEC controllers
because when the second FEC is probed the flag is already set.

This commit moves the flag inside the FEC private structure so that there
is one per controller. The flag is cleared during the probe and set on the
first call to the PHY reset function.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2701