mtd: Add cryptographic support to MTD devices.
This commit adds an "enc" MTD partition flag to mark a MTD partition as
encrypted on the command line, or an "encrypted" property when defining
partitions on the device tree.

It also modifies the mtd_read/mtd_write functions to perform data
encryption/decryption on access to encrypted partitions using the AES
symmetric block cipher and CBC (Cipher Block Chaining) mode of operation.

This means that each plaintext block is XORed with the previous ciphertext
block before being encrypted. The initialization vector (IV) for the first
block is calculated using the md5 hash of the key and a block number which
is calculated from the block offset on the physical MTD partition.

As the encryption is performed by the MTD layer any type of filesystem could
be encrypted, including UBIFS, SquashFS and JFFS2.

The encryption key used needs to be previously encrypted with the target's CAAM
as a blob by an external user space application. The application will
then store it in flash so that the kernel can retrieve it. The location of
the key is passed via either the kernel command line or the DT.

https://jira.digi.com/browse/DEL-3040

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>