ARM: dts: ccimx6ulsbc: add support for Fusion 7" touch controller
Fusion 7" display is connected to the SBC via the Video Adapter Board.
This board includes an I2C address translator chip to avoid the original
I2C address of the touch controller (0x10) conflicting with Maxim audio
chip which is at the same I2C address.
The new I2C address for the touch controller is configurable through
resistors and defaults to 0x28.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2397