ARM: dts: ccimx6ulsbc: use ENET2 MDIO bus only
The MDIO bus is shared by ENET1 and ENET2. Both interfaces use
the same pads, but may configure them for ENET1 or ENET2.
The base DTSI file for the ccimx6ulsbc was configuring these pads
for both ENET1 and ENET2 pads, which creates a conflict since
the same pads cannot be configured for both.

Since the mdio bus is defined on ENET2, let's use the MDIO ENET2
IOMUX and extend the comment for explaining that in case ENET2
is not enabled but ENET1 is, not only the mdio bus must be moved
to ENET1 but also the IOMUX for MDIO on ENET1 must be added.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3094