fsl_otp: do not disable the OTP clock
It has been observed that when the OTP registers are accessed
early during kernel initialization (during OTP driver probe)
disabling of the OTP clock produces a hang of the i.MX6UL.

Until NXP pronounces on this issue, this patch comments out the
disabling of the OTP clock.

The consumption for having the clock permanently active is
considered to be minimal.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3153