vf610-adc: add adc-ch-list and max-channel settings
This commit adds support to configure the list of channels that
will be available through the sysfs/dev interface.
The previous implementation only allowed to set the number of
available channels through the devicetree setting "num-channels",
but all the adc channels were consecutive.

The commit adds two new devicetree settings: "max-channel" and
"adc-ch-list".

max-channel: allows to define what is the highest channel number
supported. This value is platform specific. Note that this driver
is used by different SoCs (Vybrid, i.mx6UL, ...).

ad-ch-list: allows to define the list of adc channels that the driver
should enable and offer through the sysfs/dev interfaces.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>