regulators: pfuze3000: revamp the support for the pfuze3000
This commits reimplements most of the methods used by the pfuze3000
PMIC.

A significant part of the original code did not work well for this
model of PMIC (and most likely for the other supported models).
For instance, the methods t enable/disable the buck regulators were
not working.

Additionally, this commit implements methods to configure the state
of the regulators in suspend.

The code should be reviewed and generalized for all the PMIC models
supported.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>