mfd: mca: core: implement pm_ops
This commit implements proper pm_ops for the mca i2c driver.
Before, the driver was using the suspend_begin and
suspend_end hooks, to notify the MCA that the system was
entering/exiting suspend state. The problem is that these
hooks are called very early during suspend (and late during
resume) and, when the MCA enters in low power, avoids any
communication between the mfd drivers and the MCA in their
respective suspend/resume functions.

The suspend_begin/suspend_end hooks are kept here for future
use, to fully monitor the suspend/resume operation.

https://jira.digi.com/browse/DEL-3012

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>