serial: imx: force UART1 to work as DTE/no_hw_flow for CC6UL with HV < 4
UART1 on the ConnectCore 6UL is connected to the Bluetooth chip.
On SOMs with hardware version (HV) < 4, the TX and RX lines of the UART1
were crossed connected to the Bluetooth chip. In order to make the
Bluetooth work, the UART1 needed to be:
 - forced in DTE mode
 - not use HW flow control
 - use a pinctrl for DTE mode
 - configured to work at a slow speed that did not require HW flow control

This was all done by means of the Device Tree and the bluez init script.

The hardware version (HV) 4 of the SOM fixes the RX/TX lines so this patch
adapts the default Device Tree to use the correct pinctrl, default to DCE
mode, and enables HW flow control, allowing the Bluetooth to work at full
speed.
To keep compatibility of SOMs with HV < 4, the patch forces the UART1 to
DTE mode, no HW flow control, and selects the pinctrl for DTE mode.
Additionally, the RTS/CTS lines are set to GPIO IOMUX for these versions.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3402
(cherry picked from commit d745041e7a1ebe67e9c20d4cecf56f0810394b3c)