mtd: nand: nand_base: extend nand_decode_ext_id for Spansion flashes
This commit extends the decoding capabilities of the function
nand_decode_ext_id() to better support Spansion flashes that
use a larger spare area and that also include the ECC level
in the fifth byte.

This helps when the read of the ONFI parameter area fails,
a random problem observed in the ccimx6ul platform in some
units.

https://jira.digi.com/browse/CC6UL-547

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>