mtd: gpmi: fix enable_edo_mode
Some flashes support EDO modes but not the set/get feature cmd.
The enable_edo_mode function always used set/get feature cmd and
was failing with flashes that did not support that cmd, even if
the flash supported EDO modes.

This commit fixes the function checking first if the cmd is
supported.

https://jira.digi.com/browse/DEL-2878

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>