ARM: dts: ccimx6ulstarter/sbc: update pad config for clock line
Change the pad configuration of the uSDHC1 clock line to increase the DSE
to the maximum and set the speed to slow (50MHz), only for the default
pad cfg group.

This configuration seems to work better with the eMMC/uSD hardware setup
of the SBC, where the clock and cmd lines are muxed through an analog mux
what adds some extra capacitance and delays in that lines, being the clock
the more critical.

https://jira.digi.com/browse/DEL-3112

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>