ARM: imx: make the function digi_get_board_version() generic
The function was looking for the property 'digi,carrierboard,version'
only as a property of a 'digi,ccimx6' compatible machine.
Since the property is alredy Digi-custom, we can simply look for it
directly as a root node property.

While on it, also free the node's kobject and turn the pr_info message
into a pr_debug.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>