mfd: mca_cc6ul: drive MCA_FW_UPDATE GPIO low for proper CPU reset
The MCA_FW_UPDATE is a reserved GPIO whose IOMUX was being set but the line
itself was not being configured as output.
As input, the line had a voltage divisor and around 1V voltage which was
preventing the CPU from resetting when the reset button is pressed and
maintained.

This commit:
 - Creates a new optional property 'fw-update-gpio' for SOM's that have
   this line.
 - Sets the GPIO low for normal mode.
 - Creates a sysfs entry (if the property available on the DT) that can
   be used to toggle the GPIO during an MCA firmware update process.
 - On the CC6UL device tree, moves the GPIO IOMUX from the generic 'hog'
   node to the 'mca' node domain, configures the GPIO as output, and
   asserts it low by default. While on it, the pinctrl_hog is removed
   since it is now empty.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3609