Revert "ARM: dts: ccimx6ulstarter/sbc: add pinctrl pad for different uSDHC speeds"
This reverts commit b13884637aae0d8a03428388c31b5a156ba4150a.

It was verified that the CC6UL SBC does not work reliably at
an MMC clock speed of 200 MHz. Additionally, the driver seems
to allow speeds higher than 50 MHz only when the interface is
configured for 1.8 V (the UHS-I cards use it). The SBC interface
does not physically support this voltage.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3639