mfd: mca_cc6ul: i2c: use suspend_late/resume_early
Use suspend_late/resume_early in the mca_cc6ul_i2c pm_ops instead of the
regular suspend/resume callbacks to continue being functional during the
regular suspend/resume callbacks of other drivers, just in case they use
any functionality of the mca.

Make the same for the i2c-imx driver, as the mca_cc6ul uses i2c for
communication and, during the suspend/resume calls, the kernel will not
take care of that dependency.

This should be done by organizing properly the devicetree, but the imx6ul
does not react when the main dtsi file (imx6ul.dtsi) is reordered.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>