mfd: ioexp: new MCA-based I/O Expander driver.
This commit adds the necessary files for building the driver for GPIOs
and ADCs on the I/O Expander for the CC6UL SBC. The drivers are
basically the same ones for the MCA as the code has been generalized for
this purpose.

A new configuration macro (MFD_MCA_IOEXP) is added to Kconfig to control
enabling or not this feature.

Signed-off-by: Sebastian Pastor <sebastian.pastor@digi.com>