mfd: ioexp: Restore configuration on power resume
This commit adds a new optional feature to the device-tree binding of
the I/O Expander (reconfigure-on-resume) which allows restoring the
configuration of ADC, GPIOs and IRQs when leaving suspend mode.
With this property enabled the IRQs are disabled when entering suspend,
fixing the problem by which the module would not go to sleep when the
I/O Expander was present.
The DTS for the SBC have been updated to enable this property by
default. If the I/O Expander is required to be a wake-up source, then
the property should be removed but the IC would need to remain powered
in suspend, which implies other changes as well.

Related issues:
- https://jira.digi.com/browse/DEL-3961
- https://jira.digi.com/browse/DEL-3918

Signed-off-by: Sebastian Pastor <sebastian.pastor@digi.com>