ARM: dts: ccimx6ulstarter: remove Ethernet PHY gpio-regulator
The Ethernet PHY on the Starter Board is powered by a circuit
under the control of GPIO3_2. However, due to the pull-ups on
different lines the power of the PHY cannot go down to 0V which
makes this regulator unable to really control PHY power. In
fact, using it may leave the PHY partially powered which leads
to instability issues.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-3150