iio: adc: mca: add support for analog comparators
This commit modifies the common code for the MCA ADCs and adds entries
in sysfs for managing analog comparators.

There is a new configuration for enabling this feature in the kernel.

This requires the MCA firmware to be updated in order to work. That is
the reason of the heavy changes in the registers definitions and the new
IRQ for the ADC in the mca-cc6ul-i2c.c, mca-cc6ul-irq.c and mca-cc6ul-
core.c files.
Changes in mca-ioexp files are to update the names of the registers, but
since the value remains the same, it is not required to update the
firmware in the I/O Expander.

The core functionality is implemented in mca-common-adc.c which now in
the mca_adc_probe() function parses the device-tree searching for
comparator and adc channels lists. Then the ADCs channels are
initialized the same way they were before and the comparators are
treated in a different way: they use the ext_info field to add sysfs
entries to handle the thresholds, averager, comparator output, IRQ
edges, sleep behavior, etcetera. Notice that they can still be used as
regular one-shot ADCs because the raw_read attribute is exactly the
same.

Finally, the ADC IRQ from the MCA core is registered and handled. In it
the events are pushed to the /dev/ entry that an application can read to
determine which channel generated the IRQ and what was the edge, among
other information such as the timestamp.

Signed-off-by: Sebastian Pastor <sebastian.pastor@digi.com>