drivers: rtc-mca-cc6ul: fix issue with alarm wakeup
The MCA RTC alarm expires (triggers the irq) when the RTC time matches the
value programmed in the alarm register and the RTC counter increments.
This means, one second after the programmed value.

To correct this, the alarm value is adjusted when it is being written/read,
decrementing/incrementing the value by 1 second.

https://jira.digi.com/browse/CC6UL-821

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>