drivers: mmc: core: reduce overhead of CMD polling
For removable cards with the broken-cd property that do not have the
cd-gpios property, polling is done using commands. This happens once per
second for four frequencies (400, 300, 200 and 100 kHz) impacting the CPU
performance.

This patch skips certains operations (hw reset, sdio reset) in the polling
procedure if the previous conditions are met to reduce the overhead. This
should not change the behaviour for uSD cards.

https://jira.digi.com/browse/DEL-3640

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>