mtd: crypto: update default keyblob location to 'safe' partition
In dey-2.2-r2 the filesystem encryption key have been moved from
'environment' partition to 'safe' to allow bad-block handling in the
former. Now the key is stored at the first block of 'safe' partition.
U-Boot has been patched to copy the key from the old placement to the
new one if it exists.
However, to work around the case in which an older bootloader runs
with a dey-2.2-r2 kernel, the driver will try to read the key from the
previous default values.
Also, added bad-block checks before reading from the device.

Signed-off-by: Sebastian Pastor <sebastian.pastor@digi.com>
Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>

https://jira.digi.com/browse/DUB-739
https://jira.digi.com/browse/DUB-740
https://jira.digi.com/browse/DEL-6499