ARM: dts: ccimx6ulstarter/sbc: add IOMUX and entries for SPI slave
The SPI bus, when configured as master uses the chip select as GPIO because
the driver does not support native chip select.
When working as slave, the chip select line must be used as native chip
select, so we need a new IOMUX block for configuring the bus as slave.

This also adds entries (commented out by default) to configure the SPI as
slave using the spidev as driver.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-4355