drivers: fusion: reenable irq on error
When there is a touch event the interrupt handler queues the work for
reading the data from the sensor and disables the IRQ. Then this work
should reenable the IRQ after reading and processing all the data, but
it was not doing so on error.

This was causing the device to stop working after an error while handling
any interrupt. After this patch the problematic interrupt will be ignored
but the driver will keep handling future interrupts.

https://jira.digi.com/browse/DEL-4513

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>