hci_ibs: notify EIO error when communication with chip is lost
Very occassionally, under some init conditions, the BT chip is not well
initialized which leads to a corruption of the BT FW. Under this scenario,
the driver will receive an uncontrolled number of packets that are unrecognized
("Bluetooth: Unknown HCI packet type xx").
Limit the amount of printed "Unknown" messages and notice the userspace
that BT communication is lost, so user(space) needs to reattach the BT device.

https://jira.digi.com/browse/DEL-3711

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>