ARM: dts: ccimx6sbc: remove 'hdmi' label from protection IC
The SBC has an I2C HDMI transmitter port protection interface. On the
device tree this had a label 'hdmi' which was not being used.
On the v4.9 BSP, that same label has been defined for the HDMI
controller of the CPU, so this commit removes the unused one from the
I2C port protection to avoid conflicts.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>