ARM: configs: ccimx6sbc: enable i.MX IPUv3 Prefetch Resolve Engine
The PRE is a prefetch and tile resolve engine which prefetches
display data from DRAM to an internal SRAM region.
It has a single clock for configuration register access and the
functional units. A single shared interrupt is used for status and
error signaling.

The only external dependency is the SRAM region to use for the prefetch
double buffer.

Signed-off-by: Sebastian Pastor <sebastian.pastor@digi.com>

https://jira.digi.com/browse/DEL-5272