drivers: goodix: decouple irq and reset gpios
The Goodix touch controller allows the use of two optional gpios (RESET and
INT) to reset the touch controller, select the i2c address of the device and
exit the device from sleep mode.

This commit allows the use of the INT gpio without the RESET gpio. The pre-
vious implementation required that both were provided simultaneously.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>