arm: dts: ccimx6ul: add pinctrl for iomux node to setup the hog
The pinctrl_hog is a placeholder for general purpose pads whose IOMUX
can't be set from a specific driver. On the ccimx6ulstarter and ccimx6ulsbc
DTSI include files, however it was missing to set this pinctrl for the
iomux driver and so the pads in the pinctrl_hog were disregarded.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-5379