drivers: godix: fix issue in resume function when reload-fw-on-resume is enabled
This commit fixes a problem that happened when the firmware was not provided
through the firmware subystem, this means the default one was used, and the
"reload-fw-on-resume" property was set. The firmware configuration was empty
in that case and the driver failed when resending the configuration on the
resume callback.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit 0b854e9ba09a939c7414a4980ec756d7f81d70e7)
Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>