drivers: godix: fix issue in resume function when reload-fw-on-resume is enabled
This commit fixes a problem that happened when the firmware was not provided
through the firmware subystem, this means the default one was used, and the
"reload-fw-on-resume" property was set. The firmware configuration was empty
in that case and the driver failed when resending the configuration on the
resume callback.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>