drivers: godix: fix issue in resume function when reload-fw-on-resume is enabled
This commit fixes a problem that happened when the firmware was not provided
through the firmware subystem, this means the default one was used, and the
"reload-fw-on-resume" property was set. The firmware configuration was empty
in that case and the driver failed when resending the configuration on the
resume callback.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit cd24cb404520583f7b1fd4b66b6b86235c57e2ff)
(cherry picked from commit d693992bed5ae570d0d4faf7a552c284a63fdf99)