net: ethernet: assert phy reset line during suspend
This change was lost in the v4.9 merge. The power consumption of the
ConnectCore 6UL SBC Express is reduced from 30mA to 9-11mA in suspend.

Reference commit f0ddbc0153b2 ("net: ethernet: fec: add
"phy-reset-in-suspend" device tree option")

Note that the GPIO logic in the GPIO request was reverted in the original
commit and has been fixed here.

https://jira.digi.com/browse/DEL-5494

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>