ARM: dts: cc6plus: fix voltages of LDO9 and LDO10 and disable on suspend
On the CC6PLUS, these regulators go to the LGA pins to be used for
the uSDHC controllers. LDO10 is used to pull-up the SD interface of the
wireless chip.
They can be safely disabled on suspend.
While on it, fix also the min/max voltage values, according to the PMIC
datasheet.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-5436