ARM: dts: CC6UL: duplicate Goodix touch controller for address 0x5D
The Goodix touchscreen uses the IRQ line to determine the I2C address it will
use. On the 6UL, the IRQ line is floating and may differ depending on the way
the target is powered on. This prevents the touchscreen from working on some
ocassions.

By duplicating its DT entry for both possible addresses, we can assure that it
always works.

https://jira.digi.com/browse/CC6UL-953

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>