ARM: dts: cc6ul: add option to keep csi regulator on in suspend
The CSI camera device needs to be configured after being powered
up, so in order to continue normal operation after a
suspend/resume, it is given the option to keep its voltage
regulator ON in suspend mode to avoid the camera from being
powered down.

Since the CSI camera is not supported by default, this option is
not selected either.

https://jira.digi.com/browse/DEL-5803

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>