ccimx6ulsbc: fix double negation around XBee reset line
Commit 5f6c88ca70f685448894068f45ca069830595129 introduced a double
negation on the XBee reset line. As a consequence, the XBee reset works
despite the wrong definition of the GPIO polarity.

The XBee reset line is active low, so that should be the correct polarity
declared on the device tree for the GPIO.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>
(cherry picked from commit ccc9146c82e44ef185be0d5e35f3eb8cfd9c49b5)
Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>