mca-cc8x-core: increment the resume retry counter properly
Before this change, the counter was being incremented after checking its value,
causing it to go over the retry limit by one. This was making the check right
after the loop fail, since it's a strict "equals" check.

This commit makes sure the counter never goes over its intended limit and makes
the function print the correct error message and return straight away in case
of a failure.

This commit is borrowed from commit bc1d57305 that fixed this for mca-cc6ul-core

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>