mtd: gpmi-nand: reset the devices on system resume
On the ConnectCore 6UL platform, when the system goes into suspend to ram mode,
the internal 3.3V regulator is switched off. This regulator powers the nand
flash. With certain models of flashes, when the system resumes from suspend
and switches on the power supply it is required to issue a reset command for
the flash to operate properly.

https://jira.digi.com/browse/DEL-6050

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>