mtd: gpmi-nand: reset the devices on system resume
On the ConnectCore 6UL platform, when the system goes into suspend to ram mode,
the internal 3.3V regulator is switched off. This regulator powers the nand
flash. With certain models of flashes, when the system resumes from suspend
and switches on the power supply it is required to issue a reset command for
the flash to operate properly.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
Forward-ported from commit 686f27f0fb0dc2dde2108a8890fdbcc2a727641c
Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-6050
https://onedigi.atlassian.net/browse/DEL-7669