ccimx8x: reduce the amount of memory reserved for CMA to 320 MiB
In NXP's i.MX 8 QXP CPU .dtsi, 640 MiB are being reserved for CMA, causing the
OS to slow down tremendously when running on 1 GB variants. 320 MiB is enough
for the GPU driver to work fine without slowing the system down. This is also
the value that is currently used in our i.MX 6 based platforms.

https://jira.digi.com/browse/DEL-6102

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>