ARM64: dts: configure GPIOs on CC8X SBC Express Raspberry Pi header
There are many GPIOs available on the Raspberry Pi header of the
ConnectCore 8X SBC Express. Configure their IOMUX as GPIO.
GPIO is the default IOMUX for most pads but an errata of the silicon
says that some pads do not default to GPIOs, so we'd better have the
configuration explicitly on the device tree.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>